Fixed Namespaces in Exceptions
