+++ src/TwitterAds/Errors/NotFound.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Hborras\TwitterAdsSDK\TwitterAds\Errors;
+
+use Exception;
+use Hborras\TwitterAdsSDK\TwitterAdsException;
+
+class NotFound extends TwitterAdsException
+{
+    public function __construct($message, $code, Exception $previous = null, $errors)
+    {
+        parent::__construct($message, $code, null, $errors);
+    }
+}
