+++ src/TwitterAdsException.php
@@ -34,92 +34,4 @@ class TwitterAdsException extends \Exception
     }
 }
 
-class BadRequest extends TwitterAdsException
-{
-    public function __construct($message, $code, Exception $previous = null, $errors)
-    {
-        parent::__construct($message, $code, null, $errors);
-    }
-}
-
-class NotAuthorized extends TwitterAdsException
-{
-    public function __construct($message, $code, Exception $previous = null, $errors)
-    {
-        parent::__construct($message, $code, null, $errors);
-    }
-}
-
-class Forbidden extends TwitterAdsException
-{
-    public function __construct($message, $code, Exception $previous = null, $errors)
-    {
-        parent::__construct($message, $code, null, $errors);
-    }
-}
-
-class NotFound extends TwitterAdsException
-{
-    public function __construct($message, $code, Exception $previous = null, $errors)
-    {
-        parent::__construct($message, $code, null, $errors);
-    }
-}
-
-class RateLimit extends TwitterAdsException
-{
-    private $retryAfter;
-    private $resetAt;
-
-    public function __construct($message, $code, Exception $previous = null, $errors, $headers)
-    {
-        parent::__construct($message, $code, null, $errors);
-        $this->retryAfter = $headers['retry-after'];
-        $this->resetAt = $headers['rate_limit_reset'];
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getRetryAfter()
-    {
-        return $this->retryAfter;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getResetAt()
-    {
-        return $this->resetAt;
-    }
-}
-
-class ServerError extends TwitterAdsException
-{
-    public function __construct($message, $code, Exception $previous = null, $errors)
-    {
-        parent::__construct($message, $code, null, $errors);
-    }
-}
-
-class ServiceUnavailable extends TwitterAdsException
-{
-    private $retryAfter;
-
-    public function __construct($message, $code, Exception $previous = null, $errors, $headers)
-    {
-        parent::__construct($message, $code, null, $errors);
-        $this->retryAfter = $headers['retry-after'];
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getRetryAfter()
-    {
-        return $this->retryAfter;
-    }
-}
-
